Formatting clean up.

+++ ajax/observer/deleteOldResults.php
@@ -8,23 +8,25 @@
 require_once('../../db.php');
 require_once('../../ChromePhp.php');
 
-$userId = $_SESSION['user']['id'];              //fetching user id from session
+$userId = $_SESSION['user']['id'];
 $check = $_POST['check'];
 
-if ($check == 0) {       //Deletes results based on user and experiment
+//Deletes results based on user and experiment
+if ($check == 0) {
     try {
-        $stmt = $db->prepare("DELETE FROM result "
-            . "WHERE personId = '" . $userId . "' AND experimentId = :eId");
+        $stmt = $db->prepare("DELETE FROM result WHERE personId = '" . $userId . "' AND experimentId = :eId");
 
         $stmt->execute(array(':eId' => $_POST['experimentId']));
         $res = $stmt->rowCount();
 
         echo json_encode($res);
+        exit;
     } catch (PDOException $excpt) {
-
+        //
     }
 
-} else {          //Deletes all results for a given experiment
+//Deletes all results for a given experiment
+} else {
     try {
         $stmt = $db->prepare("DELETE FROM result WHERE experimentId = :eId");
 
@@ -32,8 +34,10 @@ if ($check == 0) {       //Deletes results based on user and experiment
         $res = $stmt->rowCount();
 
         echo json_encode($res);
+        exit;
     } catch (PDOException $excpt) {
-
+        //
     }
 }
+
 ?>
